realtek: 6.6: fix VLAN handling
The CPU port of realtek switches usually accept a single VID. Because
the ethernet driver has no VLAN handling as of now we can only steer
packets by setting PVID for the CPU port. VLAN handling has never been
perfect but 3 events made things worse.

- Commit a3765082164 ("rtl83xx: dsa: Do nothing when vid 0")
- Commit e691e2b302d ("rtl83xx: dsa: reset PVID to 1 instead of 0")
- Upgrade to kernel 6.6

Reasons are:

- Rejecting VID 0 closed an open shortcut
- Initialization for PVID 1 forgot to set priv->ports[port].pvid
- Kernel 6.6 does no longer clarify CPU port as untagged

To fix this call the proper initialization routine. Use a define to
make absolutely clear what the default VLAN is and where it must be
set. Additionally adapt the CPU port setting so it will always
only set a single PVID even though the kernel adds all defined VLANs
to the CPU port. Add a kernel message "enforce management VLAN xxx"
to make that clear.

REMARK! The patch always binds the lowest defined VLAN to the CPU
port. So ensure that the management IP is bound to the lowest defined
switch VLAN.

Signed-off-by: Markus Stockhausen <markus.stockhausen@gmx.de>